Added average results in addition to best results for MoneyMaker.

+++ Distribution/FireEvolverNew.cu
@@ -45,8 +45,8 @@ GPU_GLOBAL void EvolverNew(float* results, FireStarterResult* population, FireSt
 
     // Evolve the program registers for each variation.
     unsigned long long memberSeed = seed + SEED1(member);   // Unique seed for the member
-    unsigned short evolveAge = 0;
-    unsigned short bestAge = 0;
+    unsigned int evolveAge = 0;
+    unsigned int bestAge = 0;
     unsigned int registers = 0;
 
     // The first generation is initalized with random numbers.
@@ -161,8 +161,8 @@ GPU_GLOBAL void EvolverNew(float* results, FireStarterResult* population, FireSt
 
     // Evolve the program registers for each variation.
     unsigned long long memberSeed = seed + SEED1(member);   // Unique seed for the member
-    unsigned short evolveAge = 0;
-    unsigned short bestAge = 0;
+    unsigned int evolveAge = 0;
+    unsigned int bestAge = 0;
     unsigned int registers = 0;
 
     // The first generation is initalized with random numbers.
